requirements.txt
elevenlabs
faster_whisper
numpy
google-genai
PyAudio
python-dotenv
torch
